ReactTailWindFlask/requirements.txt
google-generativeai
Flask
requests
python-dotenv
flask-cors
duckduckgo-search